完善了Room,客户端制作了胜负判定

+++ Server/TankServer/TankServer/Logic/Player.cs
@@ -12,9 +12,9 @@ public class Player
     //指向ClientState
     public ClientState state;
     //坐标和旋转
-    public int x;
-    public int y;
-    public int z;
+    public float x;
+    public float y;
+    public float z;
     public float ex;
     public float ey;
     public float ez;
